requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
django-debug-toolbar==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
Pillow==10.4.0
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
Markdown==3.5
pep8-naming==0.13.2
setuptools==75.1.0
pytz==2023.3.post1
sqlparse==0.4.4
psycopg==3.1.12
psycopg-binary==3.1.12